refactor: Participant의 로직을 Name으로 분리
