+++ src/main/java/ladder/domain/participant/Participant.java
@@ -1,34 +1,13 @@
 package ladder.domain.participant;
 
-import ladder.exception.participant.InvalidNameLengthException;
-import ladder.exception.participant.NonAlphabeticNameException;
-
 public class Participant {
-    private static final int MAX_NAME_LENGTH = 5;
-    private static final String NAME_REGEX = "^[a-zA-Z]*$";
-
-    private final String name;
+    private final Name name;
 
     public Participant(String name) {
-        name = name.trim();
-        validateNameLength(name);
-        validateIsAlphabetic(name);
-        this.name = name;
-    }
-
-    private void validateNameLength(final String name) {
-        if (name.isEmpty() || name.length() > MAX_NAME_LENGTH) {
-            throw new InvalidNameLengthException();
-        }
-    }
-
-    private static void validateIsAlphabetic(final String name) {
-        if (!name.matches(NAME_REGEX)) {
-            throw new NonAlphabeticNameException();
-        }
+        this.name = new Name(name);
     }
 
     public String getName() {
-        return name;
+        return name.getValue();
     }
 }
